% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome_functions.R
\name{browser_close}
\alias{browser_close}
\title{Close a browser}
\usage{
browser_close(kill_java = T)
}
\arguments{
\item{kill_java}{Close all java instances (can affect external processes)}
}
\value{
Nothing
}
\description{
Close the 'remDr' browser. If multiple sessions are running concurrently,
kill_java must be FALSE or else all other sessions will be interrupted as well.
}
