% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome_functions.R
\name{browser_open}
\alias{browser_open}
\title{Open a web browser}
\usage{
browser_open(browser_type = "Chrome", kill_java = T, port = NA, ...)
}
\arguments{
\item{kill_java}{Close all java instances prior to running (can affect external processes).}

\item{port}{Port to use for browser. If NA, uses freeport().}

\item{...}{Additonal arguments passed on to chrome_open_browser().}
}
\value{
Nothing, but browser remDr is created in global environment.
}
\description{
Opens a Chrome or Firefox browser ('remDr') on the designated port. When multiple sessions are
running concurrently, each session must have a different port, and kill_java
must be FALSE.
}
