% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{check_in}
\alias{check_in}
\title{Log event to sandbox}
\usage{
check_in(action = "", target = "", time_limit = 60)
}
\arguments{
\item{action}{Action taken by user}

\item{target}{Local ID for target of action}

\item{time_limit}{Amount of time that can pass before another SQL call is made}
}
\description{
Log event to sandbox
}
