% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome_functions.R
\name{chrome_close_browser}
\alias{chrome_close_browser}
\title{Close a Chrome browser}
\usage{
chrome_close_browser(kill_java = T)
}
\arguments{
\item{kill_java}{Close all java instances (can affect external processes)}

\item{browser}{Browser object created by chrome_open_browser()}
}
\value{
Nothing
}
\description{
Close the 'remDr' browser. If multiple sessions are running concurrently,
kill_java must be FALSE.
}
