% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome_functions.R
\name{chrome_driver_versions}
\alias{chrome_driver_versions}
\title{Find best Chrome Driver for use}
\usage{
chrome_driver_versions(limit = 5)
}
\arguments{
\item{limit}{How many versions should be returned?}
}
\value{
The closest compatible chromedriver versions
}
\description{
Compares Chromedriver versions to Chrome version.
Orders driver versions by closeness to chrome version.
}
