% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome_functions.R
\name{chrome_open_browser}
\alias{chrome_open_browser}
\title{Open a Chrome browser}
\usage{
chrome_open_browser(
  kill_java = T,
  port = 4636L,
  chrome_ver = NA,
  print_to = NA
)
}
\arguments{
\item{kill_java}{Close all java instances (can affect external processes)}

\item{port}{Port to use for browser.}

\item{print_to}{Folder location for downloads (only needed when printing pdfs or downloading files).}

\item{chromever}{Version of Chrome to use. Leave as NA to attempt autodetection.}
}
\value{
Chrome browser object
}
\description{
Opens a Chrome browser ('remDr') on the designated port. When multiple sessions are
running concurrently, each session must have a different port, and kill_java
must be FALSE.
}
