% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome_functions.R
\name{chrome_version}
\alias{chrome_version}
\title{Determine latest Chrome version}
\usage{
chrome_version(
 
    chrome_loc = "C:\\\\Program Files (x86)\\\\Google\\\\Chrome\\\\Application\\\\chrome.exe"
)
}
\arguments{
\item{chrome_loc}{File location for Chrome.}
}
\value{
The latest compatible Chromedriver version.
}
\description{
Examines the Chrome application to identify which Chrome driver version to use.
}
