% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome_functions.R
\name{chrome_window_switch}
\alias{chrome_window_switch}
\title{Switch to a new tab}
\usage{
chrome_window_switch(windowId)
}
\arguments{
\item{windowId}{Window ID for destination tab.}
}
\value{
Nothing
}
\description{
Switch to a new tab
}
