% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\docType{data}
\name{condition_metadata}
\alias{condition_metadata}
\title{Dataframe of condition metadata}
\format{
An object of class \code{data.frame} with 289 rows and 35 columns.
}
\usage{
condition_metadata
}
\description{
Dataframe of condition metadata
}
\keyword{datasets}
