% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{create_sql_list}
\alias{create_sql_list}
\title{Convert a vector into a string list for SQL queries}
\usage{
create_sql_list(vector, quote = T, unique = T)
}
\arguments{
\item{vector}{Vector of values}

\item{quote}{T/F Should the values be surrounded by quotes?}

\item{unique}{T/F Should duplicate values be removed from the list?}
}
\description{
Convert a vector into a string list for SQL queries
}
