% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_investigation_functions.R
\name{edit_date}
\alias{edit_date}
\title{Edit a date field}
\usage{
edit_date(id, val)
}
\arguments{
\item{id}{String. HTML ID for date field.}

\item{val}{String. A date or value that can be coerced into a date.}
}
\description{
Edit a date field
}
