% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_investigation_functions.R
\name{edit_legacy}
\alias{edit_legacy}
\title{Edit a legacy field}
\usage{
edit_legacy(name, val)
}
\arguments{
\item{name}{String. HTML name attribute for  field.}

\item{val}{String. Text to enter.}
}
\description{
Edit a legacy field
}
