% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_investigation_functions.R
\name{edit_multiselect}
\alias{edit_multiselect}
\title{Edit a multiselect field}
\usage{
edit_multiselect(id, values, options = NA, pre_clear = F, other_clear = F)
}
\arguments{
\item{id}{String. HTML ID for numeric field.}

\item{values}{A vector of values to select}

\item{options}{A vector of the possible options, in order. Autodetected if not provided, but this is slower. Make sure the first value for a provided vector is "".}

\item{pre_clear}{T/F. Should current values be deselected prior to selection of values? Does not affect the other option}

\item{other_clear}{T/F. Should the other option be deselected if pre_clear is TRUE? This will delete data in the other text field.}
}
\description{
Edit a multiselect field
}
