% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{edit_numeric}
\alias{edit_numeric}
\title{Edit a numeric field}
\usage{
edit_numeric(id, val)
}
\arguments{
\item{id}{String. HTML ID for numeric field.}

\item{val}{String. A number or value that can be coerced into a number.}
}
\description{
Edit a numeric field
}
