% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_investigation_functions.R
\name{edit_quick_code}
\alias{edit_quick_code}
\title{Edit the quick code on an investigation}
\usage{
edit_quick_code(id, val, pre_clear = F, ...)
}
\arguments{
\item{id}{HTML ID for quick code field}

\item{val}{Quick code to enter}

\item{pre_clear}{T/F. Should the quick code be cleared before attempting to enter the new value?}

\item{...}{Arguments (environment) passed on to nbs_quick_code_get()}
}
\description{
Edit the quick code on an investigation
}
