% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{edit_text_field}
\alias{edit_text_field}
\title{Edit a text field}
\usage{
edit_text_field(
  id,
  val,
  id_type = "name",
  id_suffix = "_textbox",
  pre_clear = T,
  post_key = "tab"
)
}
\arguments{
\item{id}{String. HTML ID for text field. Can vary based on ID type.}

\item{val}{String. Text to enter. Can be a list of strings.}

\item{id_type}{String. HTML attribute used to find the element. Usually one of id, name, xpath, etc.}

\item{id_suffix}{String. Suffix for the element ID.}

\item{pre_clear}{T/F. Should the existing text be highlighted prior to new text being entered. If TRUE, existing text will be cleared.}

\item{post_key}{string. One of tab, enter, etc. to be pressed after the text has been entered. Tab (the default) shifts the cursor to the next element, and can help when you need the text field data read in to activate page logic.}
}
\description{
Edit a text field
}
