% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{email_draft}
\alias{email_draft}
\title{Draft and possibly send an Outlook email}
\usage{
email_draft(
  to,
  subject,
  body,
  from = NA,
  attach = c(),
  cc = c(),
  bcc = c(),
  visible = T,
  check_ooo = F,
  send = F
)
}
\arguments{
\item{to}{Main recipients}

\item{subject}{Email subject}

\item{body}{Email body. Searches for '<br>' to decide if body is html}

\item{from}{Alternate email to send on behalf of}

\item{attach}{Character vector of files to attach}

\item{cc}{Email cc}

\item{bcc}{Email bcc}

\item{visible}{Open the email in the viewer}

\item{check_ooo}{Check for and remove recipients who are out of office the entire day}

\item{send}{Send the email}
}
\value{
T/F
}
\description{
Draft and possibly send an Outlook email
}
