% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{email_send}
\alias{email_send}
\title{Send an Outlook email. Wrapper around email_draft()}
\usage{
email_send(..., visible = F, send = T)
}
\arguments{
\item{...}{options for email_draft()}

\item{visible}{Open the email in the viewer}

\item{send}{Send the email}
}
\value{
T/F
}
\description{
Send an Outlook email. Wrapper around email_draft()
}
