% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{extract_zipcode}
\alias{extract_zipcode}
\title{Pull out a 5 digit zip code from a string}
\usage{
extract_zipcode(string, comma_priority = T, map_to_state_df = NA)
}
\arguments{
\item{string}{Single string to evaluate}

\item{comma_priority}{TRUE will restrict the search to lines which have a comma,
if a comma is present. FALSE will find the first instance of 5 digits.}

\item{map_to_state_df}{Leaving as NA will return numeric zip code.
Optional data frame to map the zip codes to some other value.
A data frame with at least 2 columns; first column must be numeric.}
}
\value{
5 digit zip code or a mapped value
}
\description{
Pull out a 5 digit zip code from a string
}
