% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{find_ancestor}
\alias{find_ancestor}
\title{Find an ancestor by attribute}
\usage{
find_ancestor(x, attr, val)
}
\arguments{
\item{x}{html to search}

\item{attr}{Attribute to extract}

\item{val}{Pattern to search for}
}
\value{
xpath of attribute
}
\description{
Find an ancestor by attribute
}
