% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome_functions.R
\name{firefox_open_browser}
\alias{firefox_open_browser}
\title{Open a Firefox browser}
\usage{
firefox_open_browser(kill_java = T, port = NA)
}
\arguments{
\item{kill_java}{Close all java instances prior to running (can affect external processes).}

\item{port}{Port to use for browser. If NA, uses freeport().}
}
\value{
Firefox browser object
}
\description{
Opens a Firefox browser ('remDr') on the designated port. When multiple sessions are
running concurrently, each session must have a different port, and kill_java
must be FALSE.
}
