% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{flag_for_mark_as_reviewed}
\alias{flag_for_mark_as_reviewed}
\title{Flag labs, morbs, or case reports to be marked as reviewed}
\usage{
flag_for_mark_as_reviewed(ids, environment, program, requested_by, comments)
}
\arguments{
\item{ids}{Local IDs for labs, morb reports, or case reports}

\item{environment}{Either 'NBS Production' or 'NBS Staging'}

\item{program}{Program area that is responsible for these events or who is running the code.}

\item{requested_by}{Program area that is responsible for these events or who is running the code.}

\item{comments}{Brief description of why these are being marked as reviewed.}
}
\value{
Nothing
}
\description{
Flag labs, morbs, or case reports to be marked as reviewed
}
