% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_investigation_functions.R
\name{get_note}
\alias{get_note}
\title{Retrieve info from an note field}
\usage{
get_note(
  id,
  page_source = NA,
  row_start = 1,
  row_end = Inf,
  columns = 1:3,
  row_separator = " ^ ",
  col_separator = " | "
)
}
\arguments{
\item{id}{String. NBS question_identifier for the field.}

\item{page_source}{Page html. If NA, will pull current browser page (slower). For longer queries, use remDr$getPageSource() to pull the html once.}

\item{row_start}{Numeric. First note row to be returned. Set to Inf to start at last row.}

\item{row_end}{Numeric. Last note row to be returned. Set to Inf to end at last row.}

\item{columns}{Numeric. Which columns (1-3) to include in output}

\item{row_separator}{String. When specified alongside a col_separator, data will be combined into one string, with the col_separator between fields and the row_separator between rows. If col_separator is not NA, and row_separator is NA, a vector will be returned.}

\item{col_separator}{String. When specified column data will be combined into one string for each row, with the col_separator between fields. Set to NA to return a dataframe.}
}
\value{
string, vector, or dataframe
}
\description{
Retrieve info from an note field
}
