% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{go_to_event}
\alias{go_to_event}
\title{Go to an event}
\usage{
go_to_event(search_type, ID = NA, uid = NA, patient_page = F, verbose = T)
}
\arguments{
\item{search_type}{The type of event to search for. Can be Lab ID, Morbidity Report ID, Document ID, or Investigation ID.}

\item{ID}{Event ID. Required if this function is called from outside the patient page.}

\item{uid}{The event uid. Providing a uid may be slightly faster.}

\item{patient_page}{Boolean. Is the function called from a patient page? By default, if an ID is supplied, the home page search is used. Setting patient_page=T speeds up code where the function is called from the patient page.}

\item{verbose}{T/F. Should some extra messages be printed?}
}
\value{
T/F
}
\description{
Go to an event
}
