% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_bulk_functions.R
\name{nbs_bulk_template}
\alias{nbs_bulk_template}
\title{Create bulk update algorithm.}
\usage{
nbs_bulk_template(outfile = NA)
}
\arguments{
\item{outfile}{NA/Character vector. File name for output csv. If NA, no file is created, but a data.frame is returned.}
}
\value{
null/data.frame
}
\description{
This function creates a data.frame that is a template for nbs_bulk_updates() algorithms.
}
