% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{nbs_field_set}
\alias{nbs_field_set}
\title{Set the value of a field on the edit investigation page}
\usage{
nbs_field_set(id, value, page = NA, check_tab = F)
}
\arguments{
\item{id}{String. HTML ID or metadata label for field.}

\item{value}{String. Value to send to the field.}

\item{page}{String. The url for the edit investigation page, the condition name, or the page name. If NA, will attempt to autodetect (slower).}

\item{check_tab}{T/F. If FALSE, assumes the element is visible currently (faster). If TRUE, a check is run to see if the correct tab is selected, then selects the tab if not (slower).}
}
\description{
Given an ID, a desired value, and the page metadata, this function will determine what type of field needs to be updated, which helper function should be used, and then update the field. It can also check to be sure the correct tab is selected prior to updating the field, although organizing your code to minimize the number of checks needed is highly recommended for speed purposes.
}
