% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{nbs_file_download}
\alias{nbs_file_download}
\title{Download a file from NBS}
\usage{
nbs_file_download(
  element,
  outputname,
  url = "https://nbsproduction.tn.gov/nbs/"
)
}
\arguments{
\item{element}{HTML element for file}

\item{outputname}{Desired name for downloaded file without extension}

\item{url}{Base url for NBS instance - e.g., 'https://nbsproduction.tn.gov/nbs/'}
}
\value{
T/F if file downloaded
}
\description{
Download documents accessed through links in NBS - e.g., attachments.
}
