% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{nbs_get_page_metadata}
\alias{nbs_get_page_metadata}
\title{Retrieve page metadata for a given page}
\usage{
nbs_get_page_metadata(page = NA)
}
\arguments{
\item{page}{String. Page name found in url. If NA, the current url of the browser will be used. Currently requires ODSE access and properly formatted ODBC connection.}
}
\value{
Dataframe of page metadata
}
\description{
Retrieve page metadata for a given page
}
