% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{nbs_get_password}
\alias{nbs_get_password}
\title{Retrieve NBS password for an account}
\usage{
nbs_get_password(username)
}
\arguments{
\item{username}{NBS username}
}
\value{
NBS password
}
\description{
Retrieves a password set with nbs_set_password().
}
