% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{nbs_go_to}
\alias{nbs_go_to}
\title{Go to the page for an event or person}
\usage{
nbs_go_to(ID, prefer_morbs = F)
}
\arguments{
\item{ID}{Local ID}

\item{prefer_morbs}{T/F. Should morb reports be searched for before lab reports?}
}
\value{
T/F
}
\description{
This function accepts a wide variety of NBS local IDs and determines the correct search function to use based on the ID pattern.
}
