% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{nbs_id_convert}
\alias{nbs_id_convert}
\title{Convert a local ID into a different ID}
\usage{
nbs_id_convert(
  local_id,
  output_type = "default",
  odbc_name = "NBS_Prod",
  ods = F
)
}
\arguments{
\item{local_id}{Patient, lab, case report, or invesigation local ID}

\item{output_type}{What ID should be returned? "default" will autodetect the type.
For labs, investigations, and case reports, the default type is observation_uid,
public_health_case_uid, and nbs_document_uid, respectively. For patient local ID,
the default return value is the search ID. You can get the patient_uid by specifying that in the output_type.
Lab, case report, and person uids are only available if you specify an odbc connection
that has ODS permissions.}

\item{odbc_name}{Name of a previously created odbc connection.}

\item{ods}{T/F. Does the connection have access to the ODS? Set to F unless you need it.}
}
\value{
vector of IDs
}
\description{
Convert a local ID into a different ID
}
