% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_investigation_functions.R
\name{nbs_investigation_cancel}
\alias{nbs_investigation_cancel}
\title{Cancel out of investigation edits}
\usage{
nbs_investigation_cancel()
}
\value{
Nothing
}
\description{
Cancel out of investigation edits
}
