% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_investigation_functions.R
\name{nbs_investigation_from_patient}
\alias{nbs_investigation_from_patient}
\title{Create an investigation from a patient page}
\usage{
nbs_investigation_from_patient(
  condition,
  pre_submit_fields = list(),
  initial_data = list()
)
}
\arguments{
\item{condition}{Condition to create. Must match dropdown value exactly.}

\item{pre_submit_fields}{Only for HIV/STI. Named list with exact values for referralBasis_textbox and reviewReason_textbox.}

\item{initial_data}{Optional. Named list with data elements and their desired values.}
}
\value{
Status of investigation creation
}
\description{
Create an investigation from a patient page
}
