% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{nbs_investigation_go_to}
\alias{nbs_investigation_go_to}
\title{Go to an investigation from a patient page}
\usage{
nbs_investigation_go_to(ID)
}
\arguments{
\item{ID}{A case ID or UID. UID is slightly faster.}
}
\value{
Nothing
}
\description{
Go to an investigation from a patient page
}
