% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_investigation_functions.R
\name{nbs_investigation_race}
\alias{nbs_investigation_race}
\title{Edit the race field in an investigation}
\usage{
nbs_investigation_race(race, uncheck_others = F)
}
\arguments{
\item{race}{String with race(s) to select. Must match race option exactly!
You can specify multiple race options by separating them with commas (e.g.,"Black or African American, Other")}

\item{uncheck_others}{T/F Should races not found in race parameter be unselected? Should usually be FALSE.
Note that if TRUE, typos in the race parameter will uncheck those races. If TRUE, a blank string ("") for race will uncheck all races.}
}
\value{
None
}
\description{
Check (and possibly uncheck) various race checkboxes from the investigation edit page.
Does not check for the correct tab currently.
}
