% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_investigation_functions.R
\name{nbs_investigation_submit}
\alias{nbs_investigation_submit}
\title{Submit changes to an investigation}
\usage{
nbs_investigation_submit(check_submission = T, legacy = F)
}
\arguments{
\item{check_submission}{T/F. Should the feedback bar text be returned (page builder pages only)?}

\item{legacy}{T/F. Is this for a legacy page?}
}
\value{
Submission status for page-builder pages. NA for legacy pages.
}
\description{
Submit changes to an investigation
}
