% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_investigation_functions.R
\name{nbs_investigation_transfer}
\alias{nbs_investigation_transfer}
\title{Transfer an investigation}
\usage{
nbs_investigation_transfer(jurisdiction)
}
\arguments{
\item{jurisdiction}{String with new jurisdiction to switch to. Must match jurisdiction dropdown option exactly!}
}
\value{
None
}
\description{
Transfer an investigation to a different jurisdiction from the investigation view page.
}
