% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{nbs_is_legacy_page}
\alias{nbs_is_legacy_page}
\title{Determine if the current page is a legacy page}
\usage{
nbs_is_legacy_page()
}
\value{
T/F
}
\description{
Determine if the current page is a legacy page
}
