% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_lab_functions.R
\name{nbs_lab_associate}
\alias{nbs_lab_associate}
\title{Associate a lab to one or more investigations}
\usage{
nbs_lab_associate(case_uids, disassociate_unlisted = F)
}
\arguments{
\item{case_uids}{A vector of public health case uids. Not local IDs}

\item{disassociate_unlisted}{T/F Should cases not found in uid be unchecked?}
}
\value{
Status message bar, or 'No association' as appropriate
}
\description{
Associate a lab to one or more investigations
}
