% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_lab_functions.R
\name{nbs_lab_create}
\alias{nbs_lab_create}
\title{Create a lab from the patient page}
\usage{
nbs_lab_create(fields, inv_create = F, ...)
}
\arguments{
\item{fields}{A list of data fields to populate with information.}

\item{inv_create}{T/F Should an investigation be created from this lab?}

\item{...}{If inv_create=TRUE, ... must contain parameters needed by nbs_investigation_from_patient()}
}
\value{
Local ID of created observation
}
\description{
Create a lab from the patient page
}
