% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_lab_functions.R
\name{nbs_lab_go_to_deprecated}
\alias{nbs_lab_go_to_deprecated}
\title{Go to a lab from anywhere}
\usage{
nbs_lab_go_to_deprecated(ID = NA, uid = NA, patient_page = F, verbose = T)
}
\arguments{
\item{ID}{A lab ID. Required if this function is called from outside the patient page.}

\item{uid}{The lab uid. Providing a uid may be slightly faster.}

\item{patient_page}{Boolean. Is the function called from a patient page? By default, if an ID is supplied, the home page search is used. Setting patient_page=T speeds up code where the function is called from the patient page.}

\item{verbose}{T/F. Should some extra messages be printed?}
}
\value{
Nothing
}
\description{
Go to a lab from anywhere
}
