% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_lab_functions.R
\name{nbs_lab_mark_as_reviewed}
\alias{nbs_lab_mark_as_reviewed}
\title{Mark an event as reviewed}
\usage{
nbs_lab_mark_as_reviewed(processing_decision = NA)
}
\arguments{
\item{processing_decision}{String with a processing decision (necessary for certain program areas). Leave NA is not required.}
}
\value{
"Marked as Reviewed" or "Already marked"
}
\description{
Mark an event as reviewed
}
