% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_investigation_functions.R
\name{nbs_legacy_tab}
\alias{nbs_legacy_tab}
\title{Switch to a tab on a legacy page}
\usage{
nbs_legacy_tab(tab)
}
\arguments{
\item{tab}{Number of tab to switch to}
}
\value{
None
}
\description{
Switch to a tab on a legacy page
}
