% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{nbs_mark_as_reviewed}
\alias{nbs_mark_as_reviewed}
\title{Mark an event as reviewed}
\usage{
nbs_mark_as_reviewed()
}
\value{
"Marked as Reviewed" or "Already marked"
}
\description{
Mark an event as reviewed
}
