% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_morb_functions.R
\name{nbs_morb_create}
\alias{nbs_morb_create}
\title{Create a morbidity report from the patient page}
\usage{
nbs_morb_create(
  fields,
  inv_create = F,
  processing_decision = "Field Follow-up",
  initial_data = NA
)
}
\arguments{
\item{fields}{A list of data fields to populate with information.}

\item{inv_create}{T/F Should an investigation be created from this lab?}

\item{processing_decision}{For STI/HIV only}

\item{initial_data}{Optional. Named list with data elements and their desired values.}
}
\value{
Local ID of created morbidity report
}
\description{
Create a morbidity report from the patient page
}
