% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_investigation_functions.R
\name{nbs_notification_create}
\alias{nbs_notification_create}
\title{Create a notification}
\usage{
nbs_notification_create(comment = "Created by automated bot")
}
\arguments{
\item{comment}{Comment to add to notification}
}
\value{
None
}
\description{
Create a notification
}
