% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{nbs_page_is_legacy}
\alias{nbs_page_is_legacy}
\title{Determine if the current page is a legacy page}
\usage{
nbs_page_is_legacy()
}
\value{
T/F
}
\description{
Determine if the current page is a legacy page
}
