% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{nbs_page_metadata_get}
\alias{nbs_page_metadata_get}
\title{Retrieve page metadata for a given page}
\usage{
nbs_page_metadata_get(page = NA)
}
\arguments{
\item{page}{String. Page name found in url. If NA, the correct page is determined using the current browser page.}
}
\value{
Dataframe of page metadata
}
\description{
Retrieve page metadata for a given page
}
