% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{nbs_password_get}
\alias{nbs_password_get}
\title{Retrieve NBS password for an account}
\usage{
nbs_password_get(username)
}
\arguments{
\item{username}{NBS username}
}
\value{
NBS password
}
\description{
Retrieves a password set with nbs_password_set.
}
