% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{nbs_password_set}
\alias{nbs_password_set}
\title{Store password for an NBS account}
\usage{
nbs_password_set(username, password = "")
}
\arguments{
\item{username}{Username for account}

\item{password}{Password for account}
}
\value{
Nothing
}
\description{
Uses the keyring package to store credentials locally.
}
