% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{nbs_patient_search}
\alias{nbs_patient_search}
\title{Search for a patient by ID}
\usage{
nbs_patient_search(id)
}
\arguments{
\item{id}{The patient ID}
}
\description{
This function searches for a patient from the home page.
}
