% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{nbs_pdf_print}
\alias{nbs_pdf_print}
\title{Print the current page as a pdf}
\usage{
nbs_pdf_print(pdfname, folder = NA, legacy_print = F)
}
\arguments{
\item{pdfname}{Desired name for the pdf.}

\item{folder}{(Optional) Folder location for the pdf after download. Usually
unnecessary since pdf will download in the location specified in 'print_to'
parameter of chrome_open_browser().}

\item{legacy_print}{Print using the NBS print button. Better quality for
legacy pages, but sometimes causes errors and is slower.}
}
\value{
None
}
\description{
Prints the current NBS page as a pdf. The pdf is initially saved in the
default download directory (which must be specified in
chrome_open_browser()).
The pdf is then renamed and possibly moved to a different folder.
The default printing method pdf is not exactly equivalent to the pdf
generated by using the 'Print'
button in NBS, but performance is improved. Setting legacy_print to TRUE
will use aspects of the NBS print button, but this method is slower and
more error prone and should only be used when needed
(e.g., better formatting for printing legacy cases)
}
