% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{nbs_queue_filter}
\alias{nbs_queue_filter}
\title{Filter queue}
\usage{
nbs_queue_filter(dropdown, search_for, grepl = F, select_all = F)
}
\arguments{
\item{dropdown}{Number representing which dropdown to use (1 is far left)}

\item{search_for}{Pattern to match in dropdown options}

\item{grepl}{T/F whether to match multiple options by regex pattern, or
match one option exactly}

\item{select_all}{T/F whether to uncheck select all prior to matching}
}
\value{
None
}
\description{
Open a queue's dropdown filter and select specified elements.
When grepl = T, many results may be selected using regex pattern matching.
When select_all = T, the 'Select all' option is clicked, which reverses the
selection. If the dropdown is a text match, grepl and select_all are ignored.
}
