% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{nbs_queue_filter_supervisor}
\alias{nbs_queue_filter_supervisor}
\title{Filter queue dropdown from supervisor queue}
\usage{
nbs_queue_filter_supervisor(
  dropdown,
  search_for,
  grepl = F,
  select_all = F,
  noclick_error = !select_all
)
}
\arguments{
\item{dropdown}{Number representing which dropdown to use (1 is far left)}

\item{search_for}{Pattern to match in dropdown options}

\item{grepl}{T/F whether to match multiple options by regex pattern, or
match one option exactly}

\item{select_all}{T/F whether to uncheck select all prior to matching}

\item{noclick_error}{T/F Should an error be generated when no option is clicked (aside from select all). Defaults to the opposite of select_all.}
}
\value{
None
}
\description{
Open a queue's dropdown filter and select specified elements.
When grepl = T, many results may be selected using regex pattern matching.
When select_all = T, the 'Select all' option is clicked, which reverses the
selection. If you are attempting to select 1 or more options,
but no options are actually in the dropdown list,
all options will remain selected and be included in the shown results.
To prevent this, an error is generated by default if no option is selected.
}
