% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_queue_functions.R
\name{nbs_queue_load}
\alias{nbs_queue_load}
\title{Navigate to a queue}
\usage{
nbs_queue_load(queue)
}
\arguments{
\item{queue}{Name of the queue to load}
}
\value{
Nothing
}
\description{
Navigate to a queue
}
