% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_queue_functions.R
\name{nbs_queue_row_click}
\alias{nbs_queue_row_click}
\title{Select a link or checkbox in a queue}
\usage{
nbs_queue_row_click(row = 1, col = 1)
}
\arguments{
\item{row}{Row to select}

\item{column}{Column to select (includes checkbox column)}
}
\value{
None
}
\description{
Select a link or checkbox in a queue
}
