% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{nbs_queue_row_info}
\alias{nbs_queue_row_info}
\title{Get information from a queue row}
\usage{
nbs_queue_row_info(row = 1)
}
\arguments{
\item{row}{Row to select}
}
\value{
Named character vector of row data
}
\description{
Get information from a queue row
}
