% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{nbs_quick_code_get}
\alias{nbs_quick_code_get}
\title{Find an NBS quick code using the displayed demographic details}
\usage{
nbs_quick_code_get(full_text = "nbs bot, ESQ Tennessee")
}
\arguments{
\item{full_text}{String. The text found in an NBS profile, generated by a quick code.}
}
\value{
Quick code
}
\description{
This function reverse searches for a quick code, given the information found in an NBS profile. This requires ODSE access and a properly formatted ODBC connection.
If a code can't be found, it is set to 'nbs-bot'. Code searches and their results are stored in the qlist object, so subsequent searches for the same profile can use the archived info (faster).
}
