% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_investigation_functions.R
\name{nbs_quick_code_search}
\alias{nbs_quick_code_search}
\title{Search for an NBS entity using demographic details}
\usage{
nbs_quick_code_search(id, value, id_type = "Person", add = T)
}
\arguments{
\item{id}{The NBS element ID}

\item{value}{The demographic details, as found in a completed quick code}

\item{id_type}{What type of entity is being searched for?}
}
\value{
Quick code
}
\description{
This function uses the built in NBS search UI to find a matching entity. If no entity is found, it can add one.
}
