% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{nbs_search}
\alias{nbs_search}
\title{Search NBS for a given ID}
\usage{
nbs_search(ID_value, ID_type = "Investigation ID", verbose = T)
}
\arguments{
\item{ID_value}{Value to search for}

\item{ID_type}{Event ID type to select from dropdown}

\item{verbose}{T/F. Should a message be generated upon unsuccessful search?}
}
\value{
Nothing
}
\description{
Starting from the NBS home screen, select the correct type of event and
search by event ID.
}
