% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{nbs_tables}
\alias{nbs_tables}
\title{Pull all the tables from a patient's page}
\usage{
nbs_tables(page_type)
}
\value{
List of dataframes
}
\description{
This will pull all tables with information.
}
