% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{nbs_user_management}
\alias{nbs_user_management}
\title{Go to a user management page}
\usage{
nbs_user_management(dc_num, type = "view")
}
\arguments{
\item{dc_num}{DC number for user}

\item{type}{Do you want to 'view' or 'edit' the user?}
}
\value{
Nothing
}
\description{
Go to a user management page
}
