% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{nbs_user_set_status}
\alias{nbs_user_set_status}
\title{Set a user as active or inactive}
\usage{
nbs_user_set_status(status)
}
\arguments{
\item{status}{Desired user status. Active or Inactive.}
}
\value{
Nothing
}
\description{
Set a user as active or inactive
}
