% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\docType{data}
\name{page_metadata}
\alias{page_metadata}
\title{Dataframe of investigation page metadata}
\format{
An object of class \code{data.frame} with 18065 rows and 62 columns.
}
\usage{
page_metadata
}
\description{
Dataframe of investigation page metadata
}
\keyword{datasets}
