% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{queue_contains}
\alias{queue_contains}
\title{Filters queue columns that use a contains string search}
\usage{
queue_contains(dropdown, search_for, pagesource)
}
\arguments{
\item{dropdown}{Number representing which dropdown to use (1 is far left)}

\item{search_for}{Pattern to match in dropdown options}
}
\value{
None
}
\description{
Filters queue columns that use a contains string search
}
