% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{test_append}
\alias{test_append}
\title{Function for testing wiNBS}
\usage{
test_append(Test, Passed)
}
\arguments{
\item{Test}{String of what the test is attempting}

\item{Passed}{T/F. Did the test pass?}
}
\description{
Function for testing wiNBS
}
