% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome_functions.R
\name{window_switch}
\alias{window_switch}
\title{Switch to a new tab}
\usage{
window_switch(windowId = NA)
}
\arguments{
\item{windowId}{Window ID for destination tab.}
}
\value{
Nothing
}
\description{
Switch to a new tab
}
