% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_investigation_functions.R
\name{edit_note}
\alias{edit_note}
\title{Edit a note field}
\usage{
edit_note(id, val)
}
\arguments{
\item{id}{String. HTML ID for note field.}

\item{val}{String. Text to enter}
}
\description{
Edit a note field
}
