% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{nbs_field_get}
\alias{nbs_field_get}
\title{Retrieve field from an investigation page}
\usage{
nbs_field_get(id, page_source = NA)
}
\arguments{
\item{id}{String. NBS question_identifier for the field.}

\item{page_source}{Page html. If NA, will pull current browser page (slower). For longer queries, use remDr$getPageSource() to pull the html once.}
}
\value{
string
}
\description{
Retrieve field from an investigation page
}
