% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{nbs_investigation_go_to}
\alias{nbs_investigation_go_to}
\title{Go to an investigation from a patient page}
\usage{
nbs_investigation_go_to(ID = NA, uid = NA, environment = "NBS Production")
}
\arguments{
\item{ID}{A case ID. Required if this function is called from outside the patient page.}

\item{uid}{The case uid. Providing a uid may be slightly faster.}

\item{environment}{What environment does this occur in? Only needed if this may be called from outside NBS (requiring login) or if uid is supplied.}
}
\value{
Nothing
}
\description{
Go to an investigation from a patient page
}
