% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{nbs_investigation_submit}
\alias{nbs_investigation_submit}
\title{Submit changes to an investigation}
\usage{
nbs_investigation_submit(check_submission = T)
}
\arguments{
\item{check_submission}{T/F. Should the feedback bar text be returned?}
}
\value{
Nothing
}
\description{
Submit changes to an investigation
}
