% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_functions.R
\name{nbs_report}
\alias{nbs_report}
\title{Go to and run and NBS report}
\usage{
nbs_report(
  report,
  basic = list(id_T_T01a = Sys.Date() - 30, id_T_T01b = Sys.Date()),
  columns = NA
)
}
\arguments{
\item{report}{String. The name of the report.}

\item{basic}{Named character vector. Values to enter/select in the basic filtering page, named by their HTML ID.}

\item{columns}{NA/character or numeric vector. If NA, will export all columns. Otherwise, will export the columns named or specified via index.}
}
\description{
This function will find an NBS report by name, then enter basic filters and column selections before exporting the report results.
Will first attempt to load the home page if not already there, but this does not work from an actual report page currently.
Advanced filtering not added yet, but possible.
}
