% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome_functions.R
\name{window_switch}
\alias{window_switch}
\title{Switch to a new tab}
\usage{
window_switch(windowId = NA, close_old = F, verbose = F)
}
\arguments{
\item{windowId}{Window ID for destination tab.}

\item{close_old}{T/F. Should the old window be closed?}

\item{verbose}{T/F. Should the new window title be printed?}
}
\value{
T/F
}
\description{
Switch to a new tab
}
